Revert "port smart-reasoning agentic RAG to eino ADK (#18654)"
This reverts commit 2db8eb6c91184402e2979b6b52b77063919c7e17.
