+++ internal/agent/tool/agentic_search.go
@@ -109,11 +109,10 @@ func (a *AgenticSearchTool) InvokableRun(ctx context.Context, argumentsInJSON st
 
 	svc := GetRetrievalService()
 	tenantID := canvasTenantID(ctx)
-	// Route through canvasDatasetIDs so a trusted WithScope injection (when
-	// present) intersects the model-supplied kb_ids, preventing escalation
-	// outside the conversation's KBs. In the canvas runtime (no injection)
-	// explicit kb_ids remain authoritative, preserving prior behavior.
-	datasetIDs := canvasDatasetIDs(ctx, args.KbIDs)
+	datasetIDs := args.KbIDs
+	if len(datasetIDs) == 0 {
+		datasetIDs = canvasDatasetIDs(ctx, nil)
+	}
 	if svc == nil || tenantID == "" || len(datasetIDs) == 0 {
 		return jsonChunksEmpty(), nil
 	}
