+++ internal/service/openai_chat.go
@@ -322,12 +322,6 @@ func (s *OpenAIChatService) OpenAIChatCompletions(c *gin.Context, userID, chatID
 	if docIDsStr != "" {
 		chatKwargs["doc_ids"] = docIDsStr
 	}
-	// smart-reasoning mode switch, carried via extra_body.agent_mode.
-	if eb, ok := req.ExtraBody.(map[string]interface{}); ok {
-		if mode, hasMode := eb["agent_mode"].(string); hasMode && mode != "" {
-			chatKwargs["agent_mode"] = mode
-		}
-	}
 
 	asyncResults, asyncErr := s.pipeline.AsyncChat(ctx, userID, dialog, filteredMessages, openaiReq.Stream, chatKwargs)
 	if asyncErr != nil {
