+++ web/src/pages/next-chats/chat-stream/run-stream.ts
@@ -20,7 +20,6 @@ export type RunChatCompletionStreamParams = {
   messages: IMessage[];
   enableThinking?: string;
   enableInternet?: boolean;
-  agentMode?: string;
   llmSetting?: Variable;
 };
 
@@ -35,7 +34,6 @@ export async function runChatCompletionStream({
   messages,
   enableThinking,
   enableInternet,
-  agentMode,
   llmSetting,
 }: RunChatCompletionStreamParams): Promise<RunChatCompletionStreamResult> {
   const { beginStream, applyAnswer, endStream } = useChatStreamStore.getState();
@@ -80,7 +78,6 @@ export async function runChatCompletionStream({
         messages,
         enableThinking,
         enableInternet,
-        agentMode,
         llmSetting,
       },
       controller.signal,
