+++ web/src/pages/next-chats/chat/chat-box/next-multiple-chat-box.tsx
@@ -367,7 +367,6 @@ export function MultipleChatBox({
     async ({
       enableInternet,
       enableThinking,
-      agentMode,
     }: NextMessageInputOnPressEnterParameter) => {
       if (trim(value) === '') return;
 
@@ -375,7 +374,6 @@ export function MultipleChatBox({
         box?.({
           enableInternet,
           enableThinking,
-          agentMode,
         });
       });
     },
