+++ web/src/pages/next-chats/hooks/use-send-shared-message.ts
@@ -69,16 +69,14 @@ export const useSendSharedMessage = () => {
       id?: string,
       enableThinking?: string,
       enableInternet?: boolean,
-      agentMode?: string,
     ) => {
       const res = await send(completionUrl, {
         conversation_id: id ?? conversationId,
         quote: true,
         question: message.content,
         session_id: get(derivedMessages, '0.session_id'),
-        reasoning: agentMode ? 0 : Number(enableThinking),
+        reasoning: Number(enableThinking),
         internet: enableInternet,
-        ...(agentMode ? { agent_mode: agentMode } : {}),
         ...(chatInfo?.llm_id ? { model_name: chatInfo.llm_id } : {}),
       });
 
@@ -104,9 +102,8 @@ export const useSendSharedMessage = () => {
       message: Message,
       enableThinking?: string,
       enableInternet?: boolean,
-      agentMode?: string,
     ) => {
-      sendMessage(message, undefined, enableThinking, enableInternet, agentMode);
+      sendMessage(message, undefined, enableThinking, enableInternet);
     },
     [sendMessage],
   );
@@ -134,7 +131,6 @@ export const useSendSharedMessage = () => {
     ({
       enableThinking,
       enableInternet,
-      agentMode,
     }: NextMessageInputOnPressEnterParameter) => {
       if (trim(value) === '' || !done) return;
       const id = uuid();
@@ -154,7 +150,6 @@ export const useSendSharedMessage = () => {
           },
           enableThinking,
           enableInternet,
-          agentMode,
         );
       }
     },
